Move silx.gui._utils to silx.gui.utils._image
