+++ silx/gui/_glutils/font.py
@@ -1,7 +1,7 @@
 # coding: utf-8
 # /*##########################################################################
 #
-# Copyright (c) 2016-2017 European Synchrotron Radiation Facility
+# Copyright (c) 2016-2018 European Synchrotron Radiation Facility
 #
 # Permission is hereby granted, free of charge, to any person obtaining a copy
 # of this software and associated documentation files (the "Software"), to deal
@@ -30,11 +30,10 @@ __date__ = "13/10/2016"
 
 
 import logging
-import sys
 import numpy
-from .. import qt
-from .._utils import convertQImageToArray
 
+from ..utils._image import convertQImageToArray
+from .. import qt
 
 _logger = logging.getLogger(__name__)
